Random bug fix

+++ src/Random.hpp
@@ -6,13 +6,13 @@ namespace RT {
 namespace Random {
 
     /// @brief Returns a random uint32_t in range [0, max(uint32_t)]
-    uint32_t random_int();
+    int32_t random_int();
 
     /// @brief Sets int generator seed
     void set_int_seed(uint32_t seed);
 
     /// @brief Returns a random float in range [0.0f, 1.0f]
-    uint32_t unit_float();
+    float unit_float();
 
     /// @brief Sets float generator seed
     void set_float_seed(uint32_t seed);
